% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSpeciesDf.R
\name{cleanSpeciesDf}
\alias{cleanSpeciesDf}
\title{cleanSpeciesDf is a prototype function to parse a species lists}
\usage{
cleanSpeciesDf(inputFile)
}
\arguments{
\item{inputTXT}{a Text of the specified format}
}
\description{
cleanSpeciesDf is a function to clean a dataframe as provided by
  the enthusiasts (coleopterists) of the beetle community.

  It uses the full power of df manipulation in R and is way (about 35 times ;) faster than the parseSpeciesList approach

The resulting dataframe is a not normalized relation ( so it means a huge table with mostly redundant informations).

  It looks like:

  familiy;   genus;   subgenus; country[1];country[2];country[n];...\cr
  alpestris; STAPHYLINIDAE; Quedius; Raphirus;0;1;1;...\cr
  aereipennis; STAPHYLINIDAE; Quedius; Raphirus;1;0;1;...\cr
  .
  .
  .
}
\examples{
 ### first the basic parsing
 inputFile <- system.file("extdata", "species.chunk",   package="parseSpeciesList")
 df <- cleanSpeciesDf(inputFile)

}
\author{
Jonas Hagge
}
\references{
Löbl, I. & A. Smetana (eds): Catalogue of Palaearctic Coleoptera: \url{http://www.apollobooks.com/palaearcticcoleoptera.htm}
}

