% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFlightPlan.R
\name{colourise}
\alias{colourise}
\title{Colourise text for display in the terminal.
https://github.com/hadley/testthat/blob/717b02164def5c1f027d3a20b889dae35428b6d7/R/colour-text.r}
\usage{
colourise(text, fg = "black", bg = NULL)
}
\arguments{
\item{text}{character vector}

\item{fg}{foreground colour, defaults to white}

\item{bg}{background colour, defaults to transparent}
}
\description{
If R is not currently running in a system that supports terminal colours
the text will be returned unchanged.
}
\details{
Allowed colours are: black, blue, brown, cyan, dark gray, green, light
blue, light cyan, light gray, light green, light purple, light red,
purple, red, white, yellow
}
\examples{
print(colourise("Red", "red"))
cat(colourise("Red", "red"), "\\n")
cat(colourise("White on red", "white", "red"), "\\n")
}

