% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFlightPlan.R
\name{makeFlightPlan}
\alias{makeFlightPlan}
\title{Basic utility to create litchi compatible autonoumous flightplans for the phantom 3 drone}
\usage{
makeFlightPlan(ofN = "dji_litchi_auto_control.csv", flightAreafN = NULL,
  flightArea = NULL, terrainfollowing = FALSE, demfN = NULL,
  terrainFil = 1, flightPlanMode = "waypoints", flightAltitude = 50,
  presetFlightTask = "simple_ortho", curvesize = 0.2, rotationdir = 0,
  gimbalmode = 0, gimbalpitchangle = 0, overlap = 0.6, uavViewDir = 90,
  actiontype = NULL, actionparam = NULL)
}
\arguments{
\item{ofN}{base output filename}

\item{flightAreafN}{filename for OGR compatible extend of the flight area. Note the points are choosen in order of their digitalization.}

\item{flightArea}{if flightAreafN=NULL you may provide the coordinates by numbers c(lon1,lat1,lon2,lat2,lon3,lat3)}

\item{terrainfollowing}{if TRUE a DEM has to be provided. The function will try to download SRTM data if no DEM data is provided}

\item{demfN}{filname of the provided digital elevation model that is used for altitude flight correction}

\item{terrainFil}{\code{flightPlanMode} is equal \code{optway}  \code{terrainFil} determines the altitude difference threshold (m) that is considered  to omit terrain following waypoints from one wp to the next wp}

\item{flightPlanMode}{set the type of flightplan \code{waypoints}, \code{track}  \code{optway} are available arguments}

\item{flightAltitude}{set the flight altitude of the whole flight. it is assuming that you start the drone at the highest point of the flightarea}

\item{presetFlightTask}{argument for presetting the camera actio at each waypoint.  \code{simple_ortho} takes one picture/waypoint, \code{multi_ortho} takes 4 picture at a waypoint, two vertically down and two in forward and backward viewing direction and an angele of -60deg and \code{simple_pano} takes a 360 deg panorama picture}

\item{curvesize}{necessary litchi params you may use them but actually there is no need}

\item{rotationdir}{necessary litchi params you may use them but actually there is no need}

\item{gimbalmode}{necessary litchi params you may use them but actually there is no need}

\item{gimbalpitchangle}{necessary litchi params you may use them but actually there is no need}

\item{overlap}{overlapping ratio of the pictures}

\item{uavViewDir}{viewing angle of the camera optimal is 90 degree rotated against the flight direction}

\item{actiontype}{the actionype of the camera control c(1,1,...)}

\item{actionparam}{the parameter for the corresponding actiontype c(0,0,...)}

\item{action}{"ortho" "individual" allows an individual mission up to 8 events}

\item{altcor}{switch for altitude correction using a DEM}

\item{uavStartCoordinate}{c(lat.lon) of the planned launch position note this is important due to the altitude correction}
}
\description{
makeFlightPlan creates different types of control files for dji phantom3 uav. It makes use of the litchi software which provides an offline interface for importing csv fieles and exporting this flightplans to the litchi format as used by the phantom. Basically the tool make a not cdji/litchi cloud based flight planning for aautonous flights much easier than the available tools.Nevertheless it it ffocussing on the planning of orthophoto flights. To make it more usable some presets are defined.
}
\note{
"optway" tries to minimy the waypoints by ommiting altitude differences of optFil
      "way" samples all points as calculated by the overlap distance. optimal for autonoumous flights because the camara takes a picture at every waypoint
      "track" is optimal for flat areas and works only within the rc control range because the the uav needs a trigger signal for taking pictures
}
\examples{
### we need sp and raster ###
library(mapview)
library(raster)
library(sp)
library(geosphere)
library(rgdal)
library(tools)

## flight 50 meters above ground over a flat topography
## NOTE preset is deactivated if actiontype != NULL 
## example take a picture at each waypoint

fp<-makeFlightPlan(ofN="dji_litchi.csv", 
flightAreafN=NULL, 
flightArea=c(50.80801,8.72993,50.80590,8.731153,50.80553,8.73472), 
terrainfollowing=FALSE,
demfN=NULL,
terrainFil=1.0,
flightParams=c(flightPlanMode="waypoints",
              flightAltitude=50,
              presetFlightTask="multi_ortho",
              curvesize=0.2,
              rotationdir=0,
              gimbalmode=0,
              gimbalpitchangle=0,
              overlap=0.6,
              uavViewDir=90,
              actiontype=c(1),
              actionparam=c(0)))
## view it
mapview(fp[[1]])

## flight 50 meters above ground over a varying surface
## relief correction is enabled
## NOTE preset is used becaus actiontype == NULL 
## example take a picture at each waypoint

fp<-makeFlightPlan(ofN="dji_litchi.csv", 
flightArea=c(50.80801,8.72993,50.80590,8.731153,50.80553,8.73472), 
terrainfollowing=TRUE,
demfN="mrbiko.tif")
## view it
mapview(fp[[1]])

## flat terrain and digitize area with leafDraw

leafDraw(preset="uav")

# assuming resulting file is names "uav.json"

p<-makeFlightPlan(ofN="dji_litchi.csv",
flightAreafN = "uav.json",
terrainfollowing=FALSE)

}
\author{
Chris Reudenbach
}
\references{

}

