% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLocations.R
\name{parseCountryCode}
\alias{parseCountryCode}
\title{parseCountrycodes is a helper function to parse a modiefied iso2 country code as provided by the beatle list}
\usage{
parseCountryCode(intext, x)
}
\arguments{
\item{inLine}{a textstring of the specified format}
}
\description{
parseCountrycodes parses the assuminly iso2 country code list of the running beetle species linestring. It provides a reionally E: N: A: sorted list of country coded occurency
  .
}
\examples{
 ### examples parseCountryCode ###

 inLine='nubilus Fabricius, 1777: 213  E: AU BH BU BY CR CT CZ DE EN FI FR GB GE GR HU IR IT LA LT MC NL NR NT PL PT RO SK SL SP ST SV SZ TR UK YU "Transcaucasus"  A: IN TR'
 countryList<-parseCountryCode(inLine)

}
\author{
Chris Reudenbach
}

