% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbi.R
\name{rgbi}
\alias{rgbi}
\title{RGB indices}
\usage{
rgbi(rgb)
}
\arguments{
\item{rgb}{A \code{RasterStack} or \code{RasterBrick} object. 3
bands are mandatory (usually red, green and blue).}
}
\value{
A \code{RasterLayer} with the index calculates as:\cr
VARI (green-red)/(green+red-blue). A Visible Atmospherically Resistant Index (VARI)\cr
BI  sqrt((R**2+G**2+B*2)/3 Brightness Index\cr
SI (R-B)/(R+B) Spectra Slope Saturation Index\cr
HI (2*R-G-B)/(G-B) Primary colours Hue Index\cr
CI (R-G)/(R+G) Soil Colour Index\cr
RI R**2/(B*G**3) Redness Index\cr
NDTI (R-G)/(R+G) Normalized difference turbidity index Water\cr
VVI  (1-(r-30)/(r+30))*(1-(g-50)/(g+50))*(1-(b-1)/(b+1))\cr
aVVI  (1-(r-rx)/(r+rx))*(1-(g-gx)/(g+gx))*(1-(b-1)/(b+1))\cr
}
\description{
This function calculates various spectral indices from a RGB 
\code{Raster*} object.
}
\examples{
library(raster)
url <- "https://upload.wikimedia.org/wikipedia/commons/2/28/RGB_illumination.jpg"
dFile <- download.file(url, "Image.jpg")
img <- stack("Image.jpg") 
plotRGB(img)
rgbI <- rgbi(img)
plot(rgbI, col = gray(255:0/255))

}
\author{
Chris Reudenbach, (VVI) Florian Detsch, (VVI) Tim Appelhans
}
\references{
Planetary Habitability Laboratory (2015): Visible Vegetation Index (VVI). Available online via \url{http://phl.upr.edu/projects/visible-vegetation-index-vvi}.

Lacaux, J. P., Tourre, Y. M., Vignolles, C., Ndione, J. A., and Lafaye, M.: Classification of ponds from high-spatial resolution remote sensing: Application to Rift Valley Fever epidemics in Senegal, Remote Sens. Environ., 106, 66–74, 2007. 

Gitelson, A., et al. "Vegetation and Soil Lines in Visible Spectral Space: A Concept and Technique for Remote Estimation of Vegetation Fraction.  International Journal of Remote Sensing 23 (2002): 2537−2562. (VARI)

MADEIRA, J., BEDIDI, A., CERVELLE, B., POUGET, M. and FLAY, N., 1997, Visible spectrometric indices of hematite (Hm) and goethite (Gt) content in lateritic soils: 5490 N. Levin et al. the application of a Thematic Mapper (TM) image for soil-mapping in Brasilia, Brazil. International Journal of Remote Sensing, 18, pp. 2835–2852.
 
MATHIEU, R., POUGET, M., CERVELLE, B. and ESCADAFAL, R., 1998, Relationships between satellite-based radiometric indices simulated using laboratory reflectance data and typic soil colour of an arid environment. Remote Sensing of Environment, 66, pp. 17–28.
}

